added Fragments clean source
