--- app/src/main/java/edu/ba/baassist/TimetableActvity.java
@@ -1,59 +1,0 @@
-package edu.ba.baassist;
-
-import android.support.design.widget.NavigationView;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.LinearLayout;
-import android.widget.ListView;
-import android.widget.RelativeLayout;
-
-import java.io.BufferedInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Calendar;
-
-public class TimetableActvity extends BaseActivity {
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_timetable_actvity);
-//        RelativeLayout.LayoutParams lp =new RelativeLayout.LayoutParams(RelativeLayout.LayoutParams.MATCH_PARENT, RelativeLayout.LayoutParams.MATCH_PARENT);
-//        addContentView(findViewById(R.id.nav_timetable), lp);
-
-
-        LayoutInflater inflater = getLayoutInflater();
-        getWindow().addContentView(inflater.inflate(R.layout.activity_navigation, null, true), new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT,
-        ViewGroup.LayoutParams.MATCH_PARENT));
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
-
-    }
-
-
-    public int getActSeconds(){
-        Calendar c = Calendar.getInstance();
-        int seconds = c.get(Calendar.SECOND);
-        return seconds;
-    }
-
-//    private void getTimetable() throws IOException {
-//        URL url = new URL("http://www.android.com/");
-//        HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-//        try {
-//            InputStream in = new BufferedInputStream(urlConnection.getInputStream());
-//            readStream(in);
-//        } finally {
-//            urlConnection.disconnect();
-//        }
-
-
-
-}
