+++ app/src/main/java/edu/ba/fragments/MainFragment.java
@@ -1,0 +1,26 @@
+package edu.ba.fragments;
+
+import android.app.Fragment;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import edu.ba.baassist.R;
+
+/**
+ * Created by richa on 14.06.2017.
+ */
+
+public class MainFragment extends Fragment{
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+
+        View rootView = inflater.inflate(R.layout.fragment_main, container, false);
+
+        return rootView;
+    }
+}
