--- app/src/main/res/layout/activity_about.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="edu.ba.baassist.AboutActivity">
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="399dp"
-        android:layout_height="219dp"
-        android:text="Campus Dual App von Franz Hübner und Richard Walter"
-         />
-</android.support.constraint.ConstraintLayout>
