--- app/src/main/res/layout/activity_timetable_actvity.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="edu.ba.baassist.TimetableActvity">
-
-    <TextView
-        android:id="@+id/message"
-        android:layout_width="368dp"
-        android:layout_height="350dp"
-        android:text="Mathe bei Daniel Gembris"
-        tools:layout_editor_absoluteY="0dp"
-        tools:layout_editor_absoluteX="0dp" />
-
-
-
-</android.support.constraint.ConstraintLayout>
