+++ app/src/main/res/layout/nav_header_main.xml
@@ -10,8 +10,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:theme="@style/ThemeOverlay.AppCompat.Dark"
-    android:weightSum="1">
+    android:theme="@style/ThemeOverlay.AppCompat.Dark">
 
     <!--<ImageView-->
         <!--android:id="@+id/imageView"-->
@@ -21,20 +20,16 @@
         <!--app:srcCompat="@android:drawable/sym_def_app_icon" />-->
 
     <!--<TextView-->
-        <!--android:id="@+id/textView3"-->
         <!--android:layout_width="match_parent"-->
         <!--android:layout_height="wrap_content"-->
         <!--android:paddingTop="@dimen/nav_header_vertical_spacing"-->
-        <!--android:text="Test User"-->
-        <!--android:textAppearance="@style/TextAppearance.AppCompat.Body1"-->
-        <!--android:textColor="#FF000000"-->
-        <!--android:layout_weight="0.1" />-->
+        <!--android:text="Android Studio"-->
+        <!--android:textAppearance="@style/TextAppearance.AppCompat.Body1" />-->
 
     <!--<TextView-->
         <!--android:id="@+id/textView"-->
         <!--android:layout_width="wrap_content"-->
         <!--android:layout_height="wrap_content"-->
-        <!--android:textColor="#ffffff"-->
-        <!--android:text="Matrikel 300300" />-->
+        <!--android:text="android.studio@android.com" />-->
 
 </LinearLayout>
